NoFome

+++ NoFome/app/src/main/java/com/generatioon/nofome/Tipocesta.kt
@@ -2,6 +2,6 @@ package com.generatioon.nofome
 
 data class Tipocesta
     (var tipo: String,
-     var valor: Int) {
+     var valor: Double) {
 
 }
